feat: create separate route for forms
