+++ src/App.tsx
@@ -4,7 +4,8 @@ import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 import Layout from './components/Layout';
 import Mainpage from './pages/Mainpage';
-import Aboutpage from './pages/About';
+import Aboutpage from './pages/Aboutpage';
+import Formspage from './pages/Formspage';
 import Notfoundpage from './pages/Notfoundpage';
 
 class App extends React.Component {
@@ -14,6 +15,7 @@ class App extends React.Component {
         <Route path="/" element={<Layout />}>
           <Route index element={<Mainpage />} />
           <Route path="about" element={<Aboutpage />} />
+          <Route path="forms" element={<Formspage />} />
           <Route path="*" element={<Notfoundpage />} />
         </Route>
       </Routes>
