+++ src/components/CardsList.tsx
@@ -6,16 +6,11 @@ import './card.scss';
 
 class CardsList extends React.Component {
   render() {
-    const cards: number[] = [];
-    for (let i = 0; i < moviesJson.movies.length; i += 1) {
-      cards.push(moviesJson.movies[i].id);
-    }
+    const cards = moviesJson.movies.map((movie) => <Card key={movie.id} movie={movie} />);
 
     return (
       <section className="cards-container" role="list">
-        {cards.map((id) => (
-          <Card key={id} movie={moviesJson.movies[id - 1]} />
-        ))}
+        {cards}
       </section>
     );
   }
