+++ src/pages/Formspage.tsx
@@ -1,0 +1,9 @@
+/* eslint-disable react/prefer-stateless-function */
+import React from 'react';
+
+class Formspage extends React.Component {
+  render() {
+    return <div>Forms</div>;
+  }
+}
+export default Formspage;
